Move glyphs to the Pahawh Hmong block
The glyph for U+0030 DIGIT ZERO is, I think, intended to also be used
for U+16B50 PAHAWH HMONG DIGIT ZERO in this font, so I added a reference
to the former from the latter. Four glyphs of unknown meaning went to
the Private Use Area. The other glyphs were moved without incident.